requirements.txt
pandas==2.2.0
numpy==1.26.0
requests
python-dotenv
streamlit==1.28.0
groq==0.4.1
python-dateutil==2.8.2
plotly==5.18.0
